fix: prepared several crates for release
