+++ Cargo.toml
@@ -1,5 +1,5 @@
 [workspace.package]
-version = "0.25.5"
+version = "0.25.6"
 edition = "2024"
 authors = ["The Tari Development Community"]
 repository = "https://github.com/tari-project/tari-ootle"
@@ -78,13 +78,12 @@ tari_ootle_app_utilities = { path = "applications/tari_ootle_app_utilities" }
 
 # core crates
 tari_consensus = { path = "crates/consensus" }
-tari_consensus_types = { path = "crates/consensus_types" }
-tari_ootle_address = { path = "crates/ootle_address" }
-tari_engine = { path = "crates/engine" }
+tari_ootle_address = { path = "crates/ootle_address", version = "0.1.0" }
+tari_engine = { path = "crates/engine", version = "0.25" }
 tari_ootle_p2p = { path = "crates/p2p" }
 tari_ootle_storage = { path = "crates/storage" }
 tari_ootle_storage_sqlite = { path = "crates/storage_sqlite" }
-tari_ootle_wallet_crypto = { path = "crates/wallet/crypto" }
+tari_ootle_wallet_crypto = { path = "crates/wallet/crypto", version = "0.1" }
 tari_ootle_wallet_sdk = { path = "crates/wallet/sdk" }
 tari_ootle_wallet_sdk_services = { path = "crates/wallet/sdk_services" }
 tari_ootle_wallet_storage_sqlite = { path = "crates/wallet/storage_sqlite" }
@@ -95,11 +94,11 @@ tari_rpc_state_sync = { path = "crates/rpc_state_sync" }
 tari_state_store_rocksdb = { path = "crates/state_store_rocksdb" }
 tari_state_tree = { path = "crates/state_tree" }
 tari_template_test_tooling = { path = "crates/template_test_tooling" }
-tari_transaction_manifest = { path = "crates/transaction_manifest" }
+tari_transaction_manifest = { path = "crates/transaction_manifest", version = "0.25" }
 tari_validator_node_rpc = { path = "crates/validator_node_rpc" }
 sqlite_message_logger = { path = "networking/sqlite_message_logger" }
 tari_base_node_client = { path = "clients/base_node_client" }
-tari_indexer_client = { path = "clients/tari_indexer_client", default-features = false }
+tari_indexer_client = { path = "clients/tari_indexer_client", default-features = false, version = "0.25" }
 tari_networking = { path = "networking/core" }
 tari_rpc_framework = { path = "networking/rpc_framework" }
 tari_rpc_macros = { path = "networking/rpc_macros" }
@@ -109,6 +108,7 @@ tari_wallet_daemon_client = { path = "clients/wallet_daemon_client" }
 transaction_generator = { path = "utilities/transaction_generator" }
 
 # Dependency crates on crates.io (update versions here as well when updating the workspace version)
+tari_consensus_types = { path = "crates/consensus_types", version = "0.25" }
 tari_ootle_common_types = { path = "crates/common_types", version = "0.25" }
 tari_engine_types = { path = "crates/engine_types", version = "0.25" }
 tari_template_builtin = { path = "crates/template_builtin", version = "0.25" }
@@ -129,15 +129,15 @@ minotari_app_grpc = { git = "https://github.com/tari-project/tari.git", tag = "v
 minotari_app_utilities = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
 minotari_node_grpc_client = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
 minotari_wallet_grpc_client = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
-tari_common = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
-tari_common_sqlite = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
-tari_common_types = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
-tari_jellyfish = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
+tari_common = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1", version = "5.3.0-pre.1" }
+tari_common_sqlite = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1", version = "5.3.0-pre.1" }
+tari_common_types = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1", version = "5.3.0-pre.1" }
+tari_jellyfish = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1", version = "5.3.0-pre.1" }
 tari_metrics = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
-tari_mmr = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
+tari_mmr = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1", version = "5.3.0-pre.1" }
 tari_node_components = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
 tari_shutdown = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
-tari_sidechain = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
+tari_sidechain = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1", version = "5.3.0-pre.1" }
 tari_transaction_components = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
 minotari_console_wallet = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
 minotari_node = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
@@ -148,7 +148,7 @@ tari_p2p = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre
 
 tari_crypto = "0.22"
 tari_utilities = "0.8"
-tari_hashing = "5.2"
+tari_hashing = "5.3.0-pre.1"
 
 ## Ledger
 minotari_ledger_wallet_common = { git = "https://github.com/tari-project/tari.git", tag = "v5.3.0-pre.1" }
