+++ applications/tari_indexer/Cargo.toml
@@ -39,7 +39,6 @@ tari_shutdown = { workspace = true }
 tari_template_lib_types = { workspace = true }
 tari_ootle_transaction = { workspace = true }
 tari_validator_node_rpc = { workspace = true }
-tari_ootle_wallet_sdk = { workspace = true }
 ootle_byte_type = { workspace = true, features = ["crypto"] }
 
 anyhow = { workspace = true }
