+++ applications/tari_indexer/src/rest_api/handlers/misc.rs
@@ -31,7 +31,8 @@ pub async fn get_identity(Extension(context): Extension<HandlerContext>) -> Hand
     let response = GetIdentityResponse {
         peer_id: info.peer_id.to_string(),
         public_key: *context.public_key(),
-        public_addresses: info.listen_addrs,
+        // NOTE: we use a string here to avoid depending on multiaddr in the client
+        public_addresses: info.listen_addrs.iter().map(|addr| addr.to_string()).collect(),
     };
     Ok(context.apply_cache_control(Json(response), 1000))
 }
