+++ applications/tari_indexer/src/storage_sqlite/models/utxo.rs
@@ -6,7 +6,7 @@ use std::str::FromStr;
 use tari_engine_types::{Utxo, UtxoOutput};
 use tari_ootle_common_types::StateVersion;
 use tari_ootle_storage::{StorageError, time::PrimitiveDateTime};
-use tari_ootle_wallet_sdk::models::{UtxoBurnt, UtxoSpent, UtxoUnspent, WalletUtxoUpdate};
+use tari_indexer_client::types::{UtxoBurnt, UtxoSpent, UtxoUnspent, WalletUtxoUpdate};
 use tari_template_lib_types::{ResourceAddress, UtxoAddress, UtxoId, crypto::PedersenCommitmentBytes};
 
 use crate::storage_sqlite::{schema::utxos, serialization::deserialize_bincode};
