+++ applications/tari_indexer/src/substate_manager.rs
@@ -33,7 +33,7 @@ use tari_indexer_lib::{cached_substate_manager::CachedSubstateManager, error::In
 use tari_ootle_common_types::{Epoch, StateVersion, SubstateRequirementRef, shard::Shard, substate_type::SubstateType};
 use tari_ootle_p2p::PeerAddress;
 use tari_ootle_storage::StorageError;
-use tari_ootle_wallet_sdk::models::UtxoStateUpdateSet;
+use tari_indexer_client::types::UtxoStateUpdateSet;
 use tari_template_lib_types::{
     ResourceAddress,
     TemplateAddress,
