--- clients/tari_indexer_client/src/rest_api/mod.rs
@@ -1,4 +1,0 @@
-//   Copyright 2025 The Tari Project
-//   SPDX-License-Identifier: BSD-3-Clause
-
-pub mod types;
