--- clients/tari_indexer_client/src/rest_api/types.rs
@@ -1,6 +1,0 @@
-//   Copyright 2025 The Tari Project
-//   SPDX-License-Identifier: BSD-3-Clause
-
-use multiaddr::Multiaddr;
-use serde::{Deserialize, Serialize};
-use tari_template_lib_types::crypto::RistrettoPublicKeyBytes;
