+++ crates/template_builtin/.gitignore
@@ -1,0 +1,1 @@
+compiled/
