+++ crates/template_builtin/Cargo.toml
@@ -1,11 +1,13 @@
 [package]
 name = "tari_template_builtin"
 description = "Tari Ootle built-in templates"
-version.workspace = true
+version = "0.25.11"
 edition.workspace = true
 authors.workspace = true
 repository.workspace = true
 license.workspace = true
+# Include pre-compiled .wasm files that are gitignored
+include = ["/src", "/compiled/*.wasm", "Cargo.toml", "README.md", "LICENSE*"]
 
 [dependencies]
 tari_template_lib_types = { workspace = true }
@@ -19,6 +21,9 @@ tari_engine = { workspace = true }
 ootle_byte_type = { workspace = true, features = ["crypto"] }
 
 tari_crypto = { workspace = true }
+
+[build-dependencies]
+anyhow = { workspace = true }
 
 [features]
 default = []
