+++ crates/template_builtin/build.rs
@@ -3,12 +3,13 @@
 
 use std::{
     env,
-    error::Error,
     fs,
     io,
     path::{Path, PathBuf},
     process::Command,
 };
+
+use anyhow::Context;
 
 const TEMPLATE_BUILTINS: &[&str] = &[
     "templates/account",
@@ -17,7 +18,7 @@ const TEMPLATE_BUILTINS: &[&str] = &[
     "templates/liquidity_pool",
 ];
 
-fn main() -> Result<(), Box<dyn Error>> {
+fn main() -> anyhow::Result<()> {
     // If templates feature is disabled, do nothing
     if env::var("CARGO_FEATURE_TEMPLATES").is_err() {
         return Ok(());
@@ -33,12 +34,20 @@ fn main() -> Result<(), Box<dyn Error>> {
     if Path::new("../tari_bor").exists() {
         println!("cargo:rerun-if-changed=../tari_bor");
     }
+    let crate_path = PathBuf::from(env::var("CARGO_MANIFEST_DIR")?);
+    let compiled_path = crate_path.join("compiled");
+    fs::create_dir_all(&compiled_path).context("Create compiled directory")?;
     for template in TEMPLATE_BUILTINS {
         // we only want to rebuild if a template was added/modified
         println!("cargo:rerun-if-changed={}/src", template);
         println!("cargo:rerun-if-changed={}/Cargo.toml", template);
 
-        let template_path = env::current_dir()?.join(template);
+        let template_path = crate_path.join(template);
+        if !template_path.exists() {
+            // If the template doesn't exist, skip it. This allows us to have templates that are not included in the
+            // build context, without causing build failures.
+            continue;
+        }
 
         // compile the template into wasm
         compile_template(&template_path)?;
@@ -48,22 +57,22 @@ fn main() -> Result<(), Box<dyn Error>> {
         let wasm_path = get_compiled_wasm_path(&template_path, wasm_name);
 
         // copy the wasm binary to the root folder of the template, to be included in source control
-        let wasm_dest = template_path.join(wasm_name).with_extension("wasm");
+        let wasm_dest = compiled_path.join(wasm_name).with_extension("wasm");
         println!("cargo:rerun-if-changed={}", wasm_dest.display());
         if wasm_dest.exists() {
-            let existing_contents = fs::read(&wasm_dest)?;
-            let dest_contents = fs::read(&wasm_path)?;
+            let existing_contents = fs::read(&wasm_dest).context("Read existing_contents")?;
+            let dest_contents = fs::read(&wasm_path).context("Read dest_contents")?;
             if existing_contents == dest_contents {
                 continue;
             }
         }
-        fs::copy(wasm_path, wasm_dest)?;
+        fs::copy(wasm_path, wasm_dest).context("Copy file to dest")?;
     }
 
     Ok(())
 }
 
-fn compile_template<P: AsRef<Path>>(package_dir: P) -> Result<(), Box<dyn Error>> {
+fn compile_template<P: AsRef<Path>>(package_dir: P) -> io::Result<()> {
     let args = ["build", "--target", "wasm32-unknown-unknown", "--release"];
 
     let output = Command::new("cargo")
@@ -76,10 +85,10 @@ fn compile_template<P: AsRef<Path>>(package_dir: P) -> Result<(), Box<dyn Error>
         eprintln!("{}", String::from_utf8_lossy(&output.stdout));
         eprintln!("stderr:");
         eprintln!("{}", String::from_utf8_lossy(&output.stderr));
-        return Err(Box::new(io::Error::other(format!(
+        return Err(io::Error::other(format!(
             "Failed to compile package: {:?}",
             package_dir.as_ref(),
-        ))));
+        )));
     }
 
     Ok(())
