--- crates/template_builtin/templates/.gitignore
@@ -1,2 +1,0 @@
-# Ignore the generated wasms
-**/*.wasm
