+++ crates/template_lib/Cargo.toml
@@ -1,7 +1,7 @@
 [package]
 name = "tari_template_lib"
 description = "Tari template library provides abstrations that interface with the Tari validator engine"
-version = "0.20.0"
+version = "0.20.1"
 edition.workspace = true
 authors.workspace = true
 repository.workspace = true
