+++ crates/template_lib_types/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "tari_template_lib_types"
-version = "0.20.0"
+version = "0.20.1"
 edition.workspace = true
 authors.workspace = true
 description = "Tari template library types"
