+++ crates/wallet/sdk/Cargo.toml
@@ -20,6 +20,7 @@ tari_ootle_transaction = { workspace = true }
 tari_template_lib = { workspace = true }
 tari_template_abi = { workspace = true }
 tari_ootle_address = { workspace = true, features = ["serde"] }
+tari_indexer_client = { workspace = true, default-features = false }
 ootle_serde = { workspace = true }
 ootle_byte_type = { workspace = true }
 
@@ -53,5 +54,6 @@ ts = [
     "tari_template_lib/ts",
     "tari_ootle_transaction/ts",
     "tari_ootle_wallet_crypto/ts",
-    "tari_consensus_types/ts"
+    "tari_consensus_types/ts",
+    "tari_indexer_client/ts",
 ]
